pyproject.toml
[project]
name = "equi"
version = "0.1.0"
description = "Shakespeare RAG Chatbot with Agent Architecture"
readme = "README_agents.md"
requires-python = ">=3.9"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "weaviate-client>=4.16.5",
    "rich>=13.0.0",
    "openai>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
]


